[diagonstics] Emit system-update-committer logs over serial
This includes the system-update-committer logs in the serial logs. This
is useful because system-update-committer is responsible for deciding if
we should accept or reject an OTA. If the OTA is rejected, it might
trigger the device to reboot before the netstack has come up. When that
happens, it's difficult to capture the system logs through `fx log` or
`ffx log`.

Bug: http://b/294114858
Change-Id: Iaf0432057d97c1fc5520fa5a871b8a557b1eb311
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895079
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>