[diagnostics] Efficient histogram encoding
Histograms have used an array of Bucket {floor, ceiling, count}. This
takes a lot of space and is very inefficient to deserialize because it
creates many objects.

Histograms are changing to use parameters + array of count. If more than
50% of the counts are 0, the array of count will be sparse, and
accompanied by an array of corresponding indexes.

This change affects the ArrayContent struct in Rust, the output of
iquery, the format of the ArchiveAccessor ABI, the contents of
inspect.json in snapshots, Sampler, the Inspect Validator, and a
change detector in sdk/ctf/tests/fidl.

Fixed: 125876
Change-Id: Id8f4c0566bcad5d9c696fb37e190437c17f11a51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875538
Reviewed-by: Kendal Harland <kjharland@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Chris Phoenix <cphoenix@google.com>