[kernel][test] Convert an ASSERT to an OOPS.
When the kernel unittest framework runs a test suite, it used to
record which CPUs were online/active before the test suite started,
and again after it was finished.  Then, it would ASSERT (kernel-panic
assert, not test-assert) that sets had not changed (IOW - if a test
took a CPU offline, it turned it back on again before returning from
the test).

The trouble with this is that if a test fails and does not clean up
after itself, this ASSERT fires typically well before the DLOG drains,
and we loose the test results, making analysis tricky.  Did the test
fail, or think it succeeded?  If it failed, why did it think it
failed?

So, change this behavior.  Instead of performing this check per test
suite, instead perform it individually for every test in the suite
(helping to isolate which test may have messed things up).  In
addition, instead of ASSERTing, print a KERNEL OOPS if the
before/after online/active counts don't match, and _also_ force the
test status to failed (even if the test would have otherwise thought
it succeeded), making sure that things like CI/CQ are sure to pick up
on the failure.

Change-Id: I98456d3dd1297ddb4ccdf95fec56352f444e942d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/896005
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>