[brcmfmac] Add WPA3WrongBssid test back to auth_test without flake
WPA3WrongBssid was disabled because of its flakiness. The root cause is
that we send up two assoc results to SME in two different threads when
brcmfmac driver failed to forward an SAE frame. This change remove the
one in WlanInterface::SaeFrameTx. When SAE auth frame is sent with a
wrong bssid, an assoc result will be triggered by firmware timeout.

Bug: 131015

Multiply: auth_test: 500

Change-Id: I42723daccb35368d484c3c9567db1c1c11dc15e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893758
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>