[ld] Refactor tests
This reconsolidates the Launch object into the test fixture, so
that there is a different test fixture with the same API for each
version of the tests and the fixture provides methods to get the
test executable loaded and dynamically linked and then run,
however that happens.  Right now, that's a single test fixture
LdStartupInProcessTests but still using the TYPED_TEST machinery.
Soon this will be expanded to include separate-process versions
of the same tests, each with a different test fixture class with
which the same TYPED_TEST cases can be instantiated.  Later that
will extend to having the same tests use out-of-process dynamic
linking instead of any kind of use of the startup dynamic linker.

Bug: 130483
Change-Id: I8bb2e06293a43d5a913d5c4e93493f321a94bc28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895580
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Alex Brachet <abrachet@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>