[wifi][brcmfmac] Fix wifi metadata retrieval in DFv2
Give wifi metadata to wifi composite device instead of sdio composite
device. Forward wifi metadata from wifi device to `brcmfmac-wlanphy`
child device so that the driver that binds to `brcmfmac-wlanphy` has
access to the wifi metadata when trying to retrieve metadata. This is
all done so that the brcmfmac driver can retrieve wifi metadata
correctly in DFv2.

In DFv1, the driver for `brcmfmac-wlanphy` could get the wifi metadata
without needing this change because DFv1 gives access to all metadata of
all ancestors of the device the driver is bound to. In DFv2, a DFv1
driver only gets access to the metadata of the device it is bound to.

Add platform device fragment to wifi composite device so that its driver
can retrieve metadata. DFv2 does not support composite drivers getting
metadata from the composite device it is bound to.

Bug: 131219
Change-Id: I917735a256304dacecb0881da0c1dca3cf42d6c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891204
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>