[sys] Improve ABI compatibility errors.
Instead of talking about ABI revisions, which are difficult-to-parse
random numbers, the error message will try to talk about API levels
when it understands what API level the component was targeting.

When the ABI revision is unknown, it suggests that the problem is
likely that the OS is too old.

Bug: 129346

Change-Id: Iaa7b1f650b065d106b8296efb357c2f0b4a83651
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893945
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>