[fxtest][rewrite] Console output module
This CL implements console output for the new fxtest. The output
consists of a few lines of continually updated status strings (if
--status is true and stdout is a TTY). Formatting is controlled by
command line flags passed in to the async task.

The test runs a basic scenario for now, but will be expended to full
golden file tests when the console format is stabilized.

Change-Id: I9b09a97b688393722b03caddab2586bf5cebaab0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891789
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>
Reviewed-by: Kevin Cho <chok@google.com>