[ld] Add LdStartupCreateProcessTests
This adds a second test fixture for testing loading the startup
dynamic linker directly in a freshly-created Zircon process.  The
tests approximate what the system program loader does, but with
transparent steps that can be controlled and debugged in the test
code.  This factors out some common code shared with the Zircon
version of the LdStartupInProcessTests fixture.  It also changes
the TestProcessArgs and its uses to be somewhat more generic to
any potential bootstrap protocol.  This looks forward to adding
additional versions of this test fixture that use new bootstrap
protocols as those get developed and prototyped.

The test executables are now compiled with separate entry point
code for the in-process and separate-process cases.  The POSIX
separate-process version is built now, and will be tested later
when a POSIX separate-process test fixture is implemented.  The
Zircon separate-process version isn't quite ready to be used yet
because the test executable requires symbolic relocation for the
zx_process_exit system call from the vDSO and that is not yet
present in the startup dynamic linker.  However, local testing
has verified that the LdStartupInProcessTests case works as far
as launching dynamic linker and the executable correctly up to
the point it tries to make the system call.

Bug: 130483
Change-Id: I8f3f1647792f484926056f563bae2c633e88a487
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895581
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>