[fxfs] Fix flushing a file where last reference is from a VMO
Prior to this CL, Fxfs would panic if the last reference is from a VMO
and the file needed flushing.  The problem was that it was attempting to
spawn an async task from the thread that handles the pager port, which
isn't possible and results in a panic.

I took the opportunity to refactor how the pager works with respect to
spawning tasks and shutdown.

Change-Id: I0c44477b4be9350dae14d5addb962dcc15c5d1c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/896539
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Rui Qi Sim <rqsim@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>