[starnix] Make sure hard linked nodes resolve to same node
Hard links need to resolve to the same underlying FsNode.  Also, not all
underlying remote filesystems guarantee to provide unique IDs (and they
could span filesystems anyway), so this changes remote filesystems so
that IDs are only used if the file system is Fxfs, which is known to
have unique IDs.

A previous iteration of this CL started down the path of trying to use a
mount option, but that doesn't work for remote filesystems created via
components where there's no opportunity to specify mount options.
Still, the refactoring I did to support this seems like it might be
beneficial in future so I've left it in.

Bug: 293949025
Bug: 131807
Change-Id: I14d55b25468dad7dc8910dc3ca465f580d2afb56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/896521
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>