[netstack3] Set timer while lock held in INCOMPLETE
Currently, when a new neighbor entry is created in the INCOMPLETE state,
the NUD lock is dropped before a neighbor probe is sent and a timer is
scheduled to retransmit a probe. There is a possible race condition
where an INCOMPLETE entry transitions into another state before the
timer is scheduled, at which point it would attempt to cancel a
non-existent timer. Instead, schedule the time while the NUD lock is
held.

To do this, introduce a NudConfigContext trait that allows accessing the
device-specific NUD configuration (retransmit timer in this case) while
holding the NUD lock.

Also add a loom test demonstrating the race condition and fix.

Bug: 35185
Test: netstack3-core-test
      netstack3_core_threading_test
Change-Id: Ie882c4b1392c4258ced6122d34eba373f0f4437c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897339
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>