[starnix][signals] Prepare for different kinds of waits
Previously, the signal machinery assumed that a task could only sleep in
a Waiter. This CL abstracts this field so that we can sleep in different
ways in the future.  In a subsequent CL, we will introduce more kinds of
sleeps that are optimized for different purposes, with the first one
being optimized for futex waits.

Bug: 131974
Test: No behavior change
Change-Id: I180655e1d9a2e1392bf0edf6d69611ccaceddfbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898398
Commit-Queue: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>