[debugger][e2e] Use isolated FFX environment
Before we were relying on the global FFX daemon running in infra (and
locally) to wire up the ZX socket to the UNIX socket via `ffx debug
connect --agent-only`. This is inherently not hermetic, because we are
subject to mDNS and USB discovery issues when there are many devices
connected to an infra bot. See the linked bugs.

Multiply: host_x64/zxdb_e2e_tests
Bug: https://fxbug.dev/130461
Bug: https://fxbug.dev/287794216
Change-Id: I9aa361470c70929abffd4a3b29908c35402e37bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886475
Reviewed-by: Dwayne Slater <dwayneslater@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>