[storage][perf] Use u64 for durations
f64 was originally used for durations when we were getting fractional
nanoseconds with zx_ticks_get. After switching to using
std::time::Duration, we now only get whole nanoseconds. Keeping
everything in u64 nanoseconds is simpler to work with.

This also relaxes the requirement that durations must not be 0. VIM3 has
an A311D processor which uses an external 24MHz crystal for the clock.
This gives the clock a resolution of 41.667ns which may cause some
operations to have a 0 duration.

Fixed: 131854
Change-Id: I552d10ebc7bd6423ed8a92f3d6bb0aed06d46a20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897378
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>