[fdf][gpt] Forward metadata from emmc to gpt
Forward metadata from the `aml-sd-emmc` device all the way down to the
`gpt` device. `gpt` requires this metadata in order to correctly set up
partitions. In DFv2, this metadata is not automatically forwarded and
must be explicitly forwarded. Several devices must forward this metadata
because the hierarchy of devices from aml-sd-emmc to gpt is as follows:

1. aml-sd-emmc
2. sdmmc
3. sdmmc-mmc
4. user
5. block
6. gpt

Bug: 131933
Change-Id: Ic675f53cffd0dcb681e3b54e61a707f7160f6f36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897379
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>