[bazel_sdk] put fuchsia_clang_repository in its own file
This CL paves the way for being able to use the fuchsia_sdk_common rules
in the fuchsia_clang_repository rule. We need to pull this out of the
deps.bzl file because the deps.bzl file will load the logic which
creates the common repository. If we load the clang repository here we
will not be able to use the common repository since it will not have
been loaded yet and will lead to errors.

BUG: b/295358711
Change-Id: Ib4509caa5ad5851bcce44036577ca263fe19ff85
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899740
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>