[socket-tun] Want double buffer size for NS2
Netstack 2 doubles the value of SO_SNDBUF and SO_RCVBUF on set, and
returns the doubled value on get. Account for this in the assertions
that the setsockopt calls succeed by doubling the want value only for
NS2 (NS3 does not do this doubling).

Also replicate the clamping behavior in both stacks so that a strict
equality can be asserted.

Test: fx test tun-socket-benchmarks
Change-Id: I58b30686e53a8a075efa8344527350dd6f9b6182
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894999
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>