[ffx] Remove query from fastboot discovery
The daemon should discover USB devices passively -- there should
be no actual USB traffic until a command requiring the device
is issued, in order to avoid unnecessary crosstalk.  This code
changes the one fastboot check (the validation that the version
of the fastboot device is "0.4") from discovery to when the
device is opened in order to interface (for flashing, rebooting,
etc).

Fixes: b/293909661
Change-Id: I1a717349e5e202df69a2ae4413a6200123986525
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898539
Reviewed-by: Dwayne Slater <dwayneslater@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>