[build] Fix loadable_module() variants to copy into base root_out_dir
For the output file of a loadable_module() to be the input to
another target such as host_test_data(), there must be a copy
to the known location in the base toolchain's output directory
as is done for executables.  This doesn't come up for the
toolchains that target Fuchsia, since those generally only
consume the output files via metadata anyway.  But it does arise
with the linux_$cpu toolchains that support loadable_module()
targets that need to be used in other ways.

Bug: 130483
Change-Id: Ib1d3356adbb1c08b5d4e469f1ee3915a8f008a82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899721
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Alex Brachet <abrachet@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>