[goldfish] Reply with err if SetBufferSize fails
This CL modifies the SetBufferSize method so that it returns
`ZX_ERR_NO_MEMORY` if it can't allocate enough memory.

The previous implementation would Close the channel, sending the status
code in the epitaph. However, epitaphs are not supported in synchronous
bindings (see https://fxbug.dev/35445).

Given that the function explicitly mentions that it can fail if it can't
allocate memory, it now also returns the status code in the reply,
instead of closing the channel.

This was used in the goldfish benchmark tests[0], which were wrong
before this CL.

[0]: https://cs.opensource.google/fuchsia/fuchsia/+/main:src/graphics/tests/goldfish_benchmark/main.cc;l=110;drc=5be41348ab659c2158210c56c9a99220fcde506a

Change-Id: Ia6a3fbb817ff66367e00b0267fc2a4ec36b3e5dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900299
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>